+++ service/departamentos.js
@@ -1,5 +1,6 @@
 import { sequelize } from "../database/database.js"
 import { Departamento } from "../models/Departamento.js";
+import { Empleado } from "../models/Empleado.js";
 
 export class Service {
     constructor() { }
@@ -38,6 +39,8 @@ export class Service {
     async deleteDepartamento(id) {
         return await sequelize.transaction(async (t) => {
             const model = await Departamento.findByPk(id)
+            const exits = await Empleado.findOne({where: { idDepartamento: id }})
+            if(exits) return { status: 400, message: 'el departamento tiene miembros activos y no puede ser eliminado' }
             if (!model) return { status: 400, message: 'el departamento no se encuentra registrado' }
             await model.destroy( { transaction: t });
             return true
